+++ pandas/tests/frame/test_timezones.py
@@ -59,20 +59,6 @@ class TestDataFrameTimezones:
         # it works
         DataFrame.from_records([rec], index="begin_time")
 
-    def test_frame_tz_localize(self):
-        rng = date_range("1/1/2011", periods=100, freq="H")
-
-        df = DataFrame({"a": 1}, index=rng)
-        result = df.tz_localize("utc")
-        expected = DataFrame({"a": 1}, rng.tz_localize("UTC"))
-        assert result.index.tz.zone == "UTC"
-        tm.assert_frame_equal(result, expected)
-
-        df = df.T
-        result = df.tz_localize("utc", axis=1)
-        assert result.columns.tz.zone == "UTC"
-        tm.assert_frame_equal(result, expected.T)
-
     def test_frame_tz_convert(self):
         rng = date_range("1/1/2011", periods=200, freq="D", tz="US/Eastern")
